remove S3 wakup workaround
ChibiOS OTGv1 driver has a remote wakeup bug that prevents the device to
resume it's operation. xxx introduced a hotfix that forcefully restarted
the usb driver as a workaround. This workaround broke multiple boards
which do not use this driver / peripheral. With the update of ChibiOS
this hotfix is now obsolete.